Forbid init in non-empty directory
